Bunch of more stuff before R18
