+++ docker-compose.yml
@@ -29,8 +29,8 @@ services:
        - pgadmin:/var/lib/pgadmin
     ports:
       - "${PGADMIN_PORT:-5050}:80"
-    logging:
-      driver: none
+    #logging:
+#  driver: none
 
   redis:
     image: redis
