+++ package.json
@@ -39,8 +39,8 @@
     "@types/ms": "^0.7.31",
     "@types/multer": "^1.4.7",
     "@types/node": "^16.11.47",
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.5",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
     "@types/safe-regex": "^1.1.3",
     "@types/sinon": "^10.0.6",
     "@types/sinon-chai": "^3.2.6",
@@ -94,8 +94,9 @@
     "monaco-editor": "^0.34.1",
     "ms": "^2.1.3",
     "multer": "^1.4.5-lts.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "re-resizable": "^6.9.9",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-merge-refs": "^2.0.1",
     "react-scripts": "5.0.1",
     "reflect-metadata": "^0.1.13",
