+++ packages/lib-components/package.json
@@ -17,30 +17,30 @@
     "sort-package": "sort-package-json"
   },
   "devDependencies": {
+    "@hookform/devtools": "^4.2.2",
     "@storybook/addon-a11y": "7.0.0-alpha.38",
     "@storybook/addon-actions": "7.0.0-alpha.38",
     "@storybook/addon-docs": "7.0.0-alpha.38",
     "@storybook/addon-essentials": "7.0.0-alpha.38",
     "@storybook/addon-links": "7.0.0-alpha.38",
-    "@storybook/react": "7.0.0-alpha.38",
     "@storybook/addon-storysource": "^7.0.0-alpha.38",
     "@storybook/builder-vite": "7.0.0-alpha.38",
     "@storybook/cli": "^7.0.0-alpha.38",
+    "@storybook/react": "7.0.0-alpha.38",
     "@storybook/react-vite": "^7.0.0-alpha.38",
     "@types/jest": "^28.1.6",
     "@types/react-helmet": "^6.1.5",
     "@types/styled-components": "^5.1.25",
     "sort-package-json": "^1.53.1",
-    "typescript": "^4.5.4",
-    "@hookform/devtools": "^4.2.2"
+    "typescript": "^4.5.4"
   },
   "peerDependencies": {
-    "@testing-library/react": "^12.1.3",
+    "@testing-library/react": "^13.4.0",
     "notistack": "^3.0.0-alpha.2",
     "polished": "^4.1.3",
     "rc-slider": "^9.7.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.3.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.1",
@@ -48,10 +48,8 @@
     "web-vitals": "^2.1.2"
   },
   "dependencies": {
-    "@floating-ui/react-dom": "^1.0.0",
     "ag-grid-react": "^28.1.1",
     "framer-motion": "^6.5.1",
-    "rc-tooltip": "^5.2.2",
     "react-helmet": "^6.1.0",
     "react-hook-form": "^7.34.0",
     "simplebar-react": "^2.4.1",
